test(database): move legacy orm tests in database tests
