--- tests/Integration/ORM/Foo.php
@@ -1,17 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Tests\Tempest\Integration\ORM;
-
-use Tempest\Database\IsDatabaseModel;
-use Tempest\Database\PrimaryKey;
-
-final class Foo
-{
-    use IsDatabaseModel;
-
-    public PrimaryKey $id;
-
-    public string $bar;
-}
