--- tests/Integration/ORM/Models/AttributeTableNameModel.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace Tests\Tempest\Integration\ORM\Models;
-
-use Tempest\Database\IsDatabaseModel;
-use Tempest\Database\PrimaryKey;
-use Tempest\Database\Table;
-
-#[Table('custom_attribute_table_name')]
-final class AttributeTableNameModel
-{
-    use IsDatabaseModel;
-
-    public PrimaryKey $id;
-}
