--- tests/Integration/ORM/Models/CasterEnum.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Tests\Tempest\Integration\ORM\Models;
-
-enum CasterEnum: string
-{
-    case FOO = 'foo';
-    case BAR = 'bar';
-}
