--- tests/Integration/ORM/Models/ChildModel.php
@@ -1,28 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Tests\Tempest\Integration\ORM\Models;
-
-use Tempest\Database\HasOne;
-use Tempest\Database\IsDatabaseModel;
-use Tempest\Database\PrimaryKey;
-use Tempest\Database\Table;
-
-#[Table('child')]
-final class ChildModel
-{
-    use IsDatabaseModel;
-
-    public PrimaryKey $id;
-
-    #[HasOne]
-    public ThroughModel $through;
-
-    #[HasOne(ownerJoin: 'child2_id')]
-    public ThroughModel $through2;
-
-    public function __construct(
-        public string $name,
-    ) {}
-}
