--- tests/Integration/ORM/Models/StaticMethodTableNameModel.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace Tests\Tempest\Integration\ORM\Models;
-
-use Tempest\Database\IsDatabaseModel;
-use Tempest\Database\PrimaryKey;
-use Tempest\Database\Table;
-
-#[Table('custom_static_method_table_name')]
-final class StaticMethodTableNameModel
-{
-    use IsDatabaseModel;
-
-    public PrimaryKey $id;
-}
